fix: :bug: Fixed an issue where matlab would block when the 47th message was sent
In matlab, cogmoteGO is called through java. Using the GIN framework, cogmoteGO will print the log to standard input and output by default, but the process started by java Runtime.getRuntime ().exec () will print the buffer, which will cause the program to hang when the buffer is full. So change the log output of GIN to be turned off by default, and it will only be turned on in Debug mode.